debug: Fix printf warning
Cast always to long so the type doesn't depend on the arch.

+++ src/libAtomVM/debug.c
@@ -26,7 +26,7 @@ void debug_dump_stack(Context *ctx)
         fprintf(stderr, "DEBUG: stack: (%i) %lx\n", i, ctx->stack[i]);
     }
     fprintf(stderr, "DEBUG: \n");
-    fprintf(stderr, "DEBUG: e: %li\n", ctx->e - ctx->stack);
+    fprintf(stderr, "DEBUG: e: %li\n", (long) (ctx->e - ctx->stack));
 }
 
 void debug_print_processes_list(struct ListHead *processes)
